Split extra fields and view alter functionality into two logical classes.
